+++ cmd/study-service-app/main.go
@@ -57,6 +57,7 @@ func ensureDBIndexes(gdb *globaldb.GlobalDBService, sdb *studydb.StudyDBService)
 	for _, i := range instances {
 		sdb.CreateSurveyDefintionIndexForAllStudies(i.InstanceID)
 		sdb.CreateMessageScheduledForIndexForAllStudies(i.InstanceID)
+		sdb.CreateParticipantIDIndexForAllStudies(i.InstanceID)
 		sdb.CreateUploadedAtIndexForStudyRulesCollection(i.InstanceID)
 		// TODO: ensure other indexes as well
 	}